Move attributedText extension to Client and make sure dynamic text is used
