+++ Shared/Extensions/StringExtensions.swift
@@ -118,17 +118,4 @@ public extension String {
         // Trim and check empty string
         return (self.trimmingCharacters(in: .whitespaces) == "")
     }
-
-    /// Handles logic to make part of string bold
-    /// - Parameters:
-    ///     - boldString: the portion of the string that should be bold. Current string must already include this string.
-    ///     - font: font for entire string, part of string will be converted to bold version of this font
-    func attributedText(boldString: String, font: UIFont) -> NSAttributedString {
-        let attributedString = NSMutableAttributedString(string: self,
-                                                     attributes: [NSAttributedString.Key.font: font])
-        let boldFontAttribute: [NSAttributedString.Key: Any] = [NSAttributedString.Key.font: UIFont.boldSystemFont(ofSize: font.pointSize)]
-        let range = (self as NSString).range(of: boldString)
-        attributedString.addAttributes(boldFontAttribute, range: range)
-        return attributedString
-    }
 }
